Resolved find-bug warning
git-svn-id: http://pww.svn.pic.philips.com/DI-dev/trunk/Projects/PurAir/Android@10148 b40d9911-a802-4213-86a8-5840401a66a1

+++ PurAir/src/com/philips/cl/di/dev/pa/activity/MarkerMapFragment.java
@@ -148,7 +148,7 @@ public class MarkerMapFragment extends Fragment implements
 												// position offset
 		// baseline to the center.
 		int yPos = (int) ((canvas.getHeight() / 2) - ((paint.descent() + paint
-				.ascent()) / (float) 2));
+				.ascent()) / 2));
 		canvas.drawText(newText, xPos, yPos, paint);
 		return bm;
 	}
